requirements.txt
Flask==2.3.2
numpy==1.24.3
Pillow==9.5.0
requests==2.31.0
torch==2.0.1
torchvision==0.15.2

